+++ examples/example-exitcode/src/main.rs
@@ -28,7 +28,7 @@ fn main() {
 
     // --------- IMPORTANT ---------
     // Register `ExitCodeSetup` for the program to enable exit codes
-    program.with_setup(ExitCodeSetup::default());
+    program.with_setup(ExitCodeSetup);
     // --------- IMPORTANT ---------
 
     program.exec_and_exit();